Python/Webapps/Flask/REST APIs/3 REST-ful/2 Restful - SQL Database/2 SQLAlchemy/webapp/requirements.txt
Flask-JWT
Flask-RESTful
Flask-SQLAlchemy
